Unify shared types
